fix(admin): /password/reset の本番ビルドのプリレンダリング失敗を修正 (#347)
useSearchParams を使うクライアントコンポーネントは Next.js の本番ビルドで
Suspense 境界を要求される。境界が無く /password/reset の静的プリレンダリングが
CI の build で失敗していた(dev では顕在化しない)。フォーム本体を reset-form.tsx に
切り出し、page.tsx を Suspense でラップする構成に変更する(web 側の page/page-client
と同じ構造)。
